Varying numbers + some javadoc

+++ src/main/java/com/sharkhendrix/serialization/IdentityMapReferenceContext.java
@@ -4,42 +4,49 @@ import java.util.HashMap;
 import java.util.IdentityHashMap;
 import java.util.Map;
 
+/**
+ * Reference context that uses an {@link IdentityHashMap} to identify same
+ * instances.
+ * 
+ * @author Joannick Gardize
+ *
+ */
 public class IdentityMapReferenceContext implements ReferenceContext {
 
-	private Map<Object, Short> idByObject = new IdentityHashMap<>();
-	private Map<Short, Object> objectById = new HashMap<>();
-	private short previousId;
+    private Map<Object, Integer> idByObject = new IdentityHashMap<>();
+    private Map<Integer, Object> objectById = new HashMap<>();
+    private int previousId;
 
-	@Override
-	public void resetWriteContext() {
-		idByObject.clear();
-		previousId = 0;
-	}
+    @Override
+    public void resetWriteContext() {
+        idByObject.clear();
+        previousId = 0;
+    }
 
-	@Override
-	public void resetReadContext() {
-		objectById.clear();
-	}
+    @Override
+    public void resetReadContext() {
+        objectById.clear();
+    }
 
-	@Override
-	public short retrieve(Object o) {
-		return idByObject.getOrDefault(o, (short) -1);
-	}
+    @Override
+    public int retrieve(Object o) {
+        return idByObject.getOrDefault(o, -1);
+    }
 
-	@Override
-	public short store(Object o) {
-		short id = ++previousId;
-		idByObject.put(o, previousId);
-		return id;
-	}
+    @Override
+    public int store(Object o) {
+        int id = ++previousId;
+        idByObject.put(o, previousId);
+        return id;
+    }
 
-	@Override
-	public void store(short id, Object o) {
-		objectById.put(id, o);
-	}
+    @Override
+    public void store(int id, Object o) {
+        objectById.put(id, o);
+    }
 
-	@Override
-	public Object retrieve(short id) {
-		return objectById.get(id);
-	}
+    @Override
+    public Object retrieve(int id) {
+        return objectById.get(id);
+    }
 }
